Consolidate select geography components
